feat: Update dev_ci.yml to use safebase_db instead of postgres and update port
This commit updates the dev_ci.yml file in the .github/workflows directory to use the service name "safebase_db" instead of "postgres" for the PostgreSQL container. It also updates the port mapping from 5432 to 5434. Additionally, it adds a volume mapping to mount the safebase.sql file from the repository into the container.